Change ID's from Integer to Long

+++ energymanagementsystem/src/main/java/rs/energymanagementsystem/energymanagementsystem/entities/AlarmData.java
@@ -19,7 +19,7 @@ public class AlarmData {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private Integer alarmId;
+    private Long alarmId;
 
     @Column
     private Integer alarmCode;
