+++ energymanagementsystem/src/main/java/rs/energymanagementsystem/energymanagementsystem/entities/ElectricityLiveData.java
@@ -16,7 +16,7 @@ public class ElectricityLiveData {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private Integer liveDataId;
+    private Long liveDataId;
 
     @Column
     private Integer consumption;
